chore(db-iam): migrate RBAC/FGA tables from db-sample to db-iam
