+++ apps/sample-api/.env.json
@@ -16,9 +16,12 @@
   "WS_OPTIONS": {},
   // Drizzle connection — set DATABASE_URL in .env
   "DRIZZLE_PG": {},
+  // Drizzle connection for db-iam — set DRIZZLE_PG_IAM_URL in .env (port 5433)
+  "DRIZZLE_PG_IAM": {},
   // service config types available: drizzle, knex, keyv, redis, ws
   "SERVICES_CONFIG": {
     "drizzle1": { "type": "drizzle", "options": "DRIZZLE_PG" },
+    "drizzle2": { "type": "drizzle", "options": "DRIZZLE_PG_IAM" },
     // "knex1": { "type": "knex", "options": "KNEXFILE_PG" }, // legacy — keep for t4t if needed
     // "knex0": { "type": "knex", "options": "KNEXFILE_MYSQL" },
     "keyv": { "type": "keyv", "options": "KEYV_CACHE" },
@@ -230,6 +233,7 @@
   // When enabled, createToken embeds active_tenant + tenants map in the JWT.
   // Can be used alongside FGA (coarse-grained RBAC + fine-grained FGA).
   "RBAC_CONFIG": {
-    "enabled": false
+    "enabled": false,
+    "serviceName": "drizzle2"
   }
 }
