+++ scripts/dbdeploy/db-iam/drizzle.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
   dialect: 'postgresql',
-  schema: './schema.ts',
+  schema: '../../../common/compiled/node/services/db-iam/schema.ts',
   out: './drizzle',
   dbCredentials: {
     url: process.env.DATABASE_URL ?? 'postgresql://localhost:5433/db_iam',
